Move decode_event function to utils/filters
