+++ raiden/utils/filters.py
@@ -1,5 +1,11 @@
 from typing import Dict
 
+from eth_utils import (
+    decode_hex,
+    event_abi_to_log_topic,
+)
+from web3.utils.abi import filter_by_type
+from web3.utils.events import get_event_data
 from eth_utils import to_checksum_address
 from web3.utils.filters import construct_event_filter_params
 from raiden_contracts.contract_manager import CONTRACT_MANAGER
@@ -59,3 +65,27 @@ def get_filter_args_for_all_events_from_channel(
     event_filter_params['topics'] = [None, event_filter_params['topics'][1]]
 
     return event_filter_params
+
+
+def decode_event(abi: Dict, log: Dict):
+    """ Helper function to unpack event data using a provided ABI
+
+    Args:
+        abi: The ABI of the contract, not the ABI of the event
+        log: The raw event data
+
+    Returns:
+        The decoded event
+    """
+    if isinstance(log['topics'][0], str):
+        log['topics'][0] = decode_hex(log['topics'][0])
+    elif isinstance(log['topics'][0], int):
+        log['topics'][0] = decode_hex(hex(log['topics'][0]))
+    event_id = log['topics'][0]
+    events = filter_by_type('event', abi)
+    topic_to_event_abi = {
+        event_abi_to_log_topic(event_abi): event_abi
+        for event_abi in events
+    }
+    event_abi = topic_to_event_abi[event_id]
+    return get_event_data(event_abi, log)
