+++ requirements-dev.txt
@@ -27,4 +27,4 @@ releases==1.5.0
 # Release
 bumpversion==0.5.3
 
-eth-tester[py-evm]
+eth-tester[py-evm]==0.1.0b28
